optimize duplication for poolcard
